flask_server/requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
bidict==0.22.1
blinker==1.6.3
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
Flask==3.0.0
Flask-SocketIO==5.3.6
frozenlist==1.4.0
gTTS==2.4.0
gunicorn==21.2.0
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
openai==0.28.1
packaging==23.2
python-engineio==4.7.1
python-socketio==5.9.0
requests==2.31.0
simple-websocket==1.0.0
SpeechRecognition==3.10.0
tqdm==4.66.1
urllib3==1.26.14 
Werkzeug==3.0.0
wsproto==1.2.0
yarl==1.9.2
flask-cors==4.0.1
pyyaml==6.0.2
python-dotenv==1.0.1
boto3==1.35.49
SQLAlchemy==1.4.46
Flask-SQLAlchemy==3.0.0
pymysql==1.0.3
cryptography==41.0.3
flask_login==0.6.3

